refactor(painter-dom): extract note frame attribute helper
Wrap the painter read-only contenteditable assignment into
applyNoteStoryFrameAttributes so the two fragment-frame call sites share
one helper, and move its coverage into a focused test alongside the new
module.
