+++ packages/layout-engine/painters/dom/src/notes/frame.ts
@@ -1,0 +1,7 @@
+import { shouldApplyPainterReadOnly } from './story.js';
+
+export const applyNoteStoryFrameAttributes = (el: HTMLElement, blockId: string | undefined): void => {
+  if (shouldApplyPainterReadOnly(blockId)) {
+    el.setAttribute('contenteditable', 'false');
+  }
+};
